requirements.txt
discord.py
Red-DiscordBot